+++ beast-of-argh/CHANGELOG.md
@@ -6,6 +6,7 @@
 * Revert to consul json config file until consul fixed start with hcl
 * Update loki and promtail to 2.7.5
 * Update loki configuration file, permissions, sysrc entries
+* Update grafana for grafana.ini and copy provisions dir from /usr/local/etc/grafana now
 
 ---